check trashbin after deletion in UI tests
