+++ tests/ui/features/lib/TrashbinPage.php
@@ -1,0 +1,38 @@
+<?php
+/**
+ * ownCloud
+ *
+ * @author Artur Neumann <artur@jankaritech.com>
+ * @copyright 2017 Artur Neumann artur@jankaritech.com
+ *
+ * This code is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License,
+ * as published by the Free Software Foundation;
+ * either version 3 of the License, or any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>
+ *
+ */
+
+namespace Page;
+
+/**
+ * Trashbin page.
+ */
+class TrashbinPage extends FilesPageBasic {
+
+	/**
+	 *
+	 * @var string $path
+	 */
+	protected $path = '/index.php/apps/files/?view=trashbin';
+	protected $fileNamesXpath = "//span[contains(@class,'nametext')]";
+	protected $fileNameMatchXpath = "//span[contains(@class,'nametext') and .=%s]";
+	protected $fileListXpath = ".//div[@id='app-content-trashbin']//tbody[@id='fileList']";
+}
